Rewrite cache() tests using Flight
cache() on the client is not yet stable. It's only stable in Server
Components. This rewrites the tests to use the Flight renderer instead
of React Noop.

I also split the cache() tests into a separate module from the tests
for the <Cache> element, which is a different (experimental) feature.
